tjpr_api_diario/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pymongo>=4.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0